% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gets.R
\name{dnb_search}
\alias{dnb_search}
\title{Search the DNB catalogue - simple search}
\source{
\url{http://www.dnb.de/EN/Service/DigitaleDienste/SRU/sru_node.html}
}
\usage{
dnb_search(title, author, year, publisher, keyword, type, language,
  limit = 10, print = FALSE)
}
\arguments{
\item{title}{the title (including subtitle, short title, volume title, etc.); optional single string value or vector of strings.}

\item{author}{the author(s); optional single string value or vector of strings.}

\item{year}{the year of publishing; optional single integer value or vector of integers.}

\item{publisher}{the publisher (publisher name and/or location); optional single string value or vector of strings.}

\item{keyword}{one or a set of keywords describing the work (subjects, persons, locations, organisations, etc.); optional single string value or vector of strings.}

\item{type}{the type of publication (optional), one or a vector of \code{articles}, \code{manuscript}, \code{biographicaldoc}, \code{letters}, \code{bequest}, \code{collections}, \code{books}, \code{brailles}, \code{maps}, \code{discs}, \code{dissertations}, \code{online}, \code{films}, \code{microfiches}, \code{multimedia}, \code{music}, \code{scores}, \code{serials}, \code{persons}, \code{subjects}, \code{corperations}, \code{works}, \code{events}, \code{geographics}.}

\item{language}{the language of the work by ISO 639-2/B code (\url{http://www.dnb.de/SharedDocs/Downloads/DE/DNB/standardisierung/inhaltserschliessung/sprachenCodesEnglisch.pdf?__blob=publicationFile}); single string value or vector of strings.}

\item{limit}{number (and starting point) of results returned; single integer value (number of results, 1--100), vector of two integer values (number of and starting point for list of results, >=1) or \code{"all"} for a complete list of results.}

\item{print}{if \code{TRUE} the search results are printed (default is \code{FALSE}).}
}
\value{
A list of results with metadata.
}
\description{
\code{dnb_search} exposes a search in the DNB catalogue.
}
\details{
to do
}
\examples{
\dontrun{
}
}

