% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{n_rec}
\alias{n_rec}
\title{Number of records}
\usage{
n_rec(dnb_obj)
}
\arguments{
\item{dnb_obj}{the DNB-search object returned \code{\link{dnb_search}} or \code{\link{dnb_advanced}}.}
}
\value{
Number of records found.
}
\description{
\code{n_rec} returns the number of items in a list of records returned by a DNB-search.
}
\examples{
\dontrun{
dnb.srch <- dnb_search(title="katze")
n_rec(dnb.srch)
}
}
