% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print_query}
\alias{print_query}
\title{Print search query}
\usage{
print_query(dnb_obj)
}
\arguments{
\item{dnb_obj}{the DNB-serch object returned by \code{\link{dnb_search}} or \code{\link{dnb_advanced}}.}
}
\value{
Query string.
}
\description{
\code{print_query} prints out the query used for a DNB-search request.
}
\examples{
\dontrun{
dnb.srch <- dnb_search(title="katze")
print_query(dnb.srch)
}
}

