% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_token}
\alias{save_token}
\title{Save token to file}
\usage{
save_token(token = dnb_token(), path = paste(normalizePath("~/"),
  ".Renviron", sep = "/"))
}
\arguments{
\item{token}{the personal DNB token as string.}

\item{path}{the path to the file where the token is stored. Default is the .Renvion file in the users home directory.}
}
\description{
\code{save_token} saves the DNB token to file, so the user does not to enter it for each R session.
}
\note{
If an environment variable named "DNB_TOKEN" is found in the file, it is updated to the given token.
}
\examples{
\dontrun{
save_token(token="YOUR_TOKEN_HERE")
}
}
