% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gets.R
\name{dnb_search}
\alias{dnb_search}
\title{Search the DNB catalogue - simple search}
\source{
\url{http://www.dnb.de/EN/Service/DigitaleDienste/SRU/sru_node.html}
}
\usage{
dnb_search(title, author, year, publisher, keyword, type, language,
  print = FALSE)
}
\arguments{
\item{title}{the title (including subtitle, short title, volume title, etc.); single string value or vector of strings.}

\item{author}{the author(s); single string value or vector of strings.}

\item{year}{the year of publishing; single integer value or vector of integers.}

\item{publisher}{the publisher (publisher name and/or location); single string value or vector of strings.}

\item{keyword}{one or a set of keywords describing the work (subjects, persons, locations, organisations, etc.); single string value or vector of strings.}

\item{type}{the type of publication, one or a vector of articles, manuscript, biographicaldoc, letters, bequest, collections, books, brailles, maps, discs, dissertations, online, films, microfiches, multimedia, music, scores, serials, persons, subjects, corperations, works, events, geographics.}

\item{language}{the language of the work by ISO 639-2/B code (\url{http://www.dnb.de/SharedDocs/Downloads/DE/DNB/standardisierung/inhaltserschliessung/sprachenCodesEnglisch.pdf?__blob=publicationFile}); single string value or vector of strings.}

\item{print}{if \code{TRUE} the search results are printed (default is \code{FALSE}).}
}
\value{
A list of results with metadata.
}
\description{
\code{dnb_search} exposes a search in the DNB catalogue.
}
\details{
to do
}
\examples{
\dontrun{
}
}

